Repaired the function symbol generator.
The function symbol generator -- when generating an arbitrary unique
symbol -- calculated in a wrong way what its basic string without
trailing numbers was. Besides that it did not request the
function_symbol generator to properly register this string.

This caused a pbes_chaining test to fail.

It has now been resolved.