Calculate a smaller, yet still correct, strategy
In the symbolic solver we overapproximate the strategies to ensure
we retain at least all dependencies that are relevant when performing
the second pass. In `safe_control_predecessors`, the strategy would
be extended for all predecessors of a set U, even if we already had
a strategy for these predecessors (i.e., those states already in U).
We now restrict this to only the *new* predecessors.