Added a new tool lpsparvalues
The tool lpsparvalues takes an lps and tries to enumerate all
possible elements of its parameter domains. By default it reports
the product of the sizes of the individual parameter domains.
Using the verbose flag it also lists the domains.

This tool may not terminate. Sometimes adding some extra rewrite
rules or increasing the iteration limit using the flag --qlimit
may help. Also applying lpssuminst with the -f flag may have
good results.

As it stands the tools is not particularly optimised yet.