Repaired an issue in the jitty rewriter.
The Jitty rewriter had a -- programming shortcut -- where
due to laziness of the programmer a term would be rewritten multiple
times if this term was the of the shape f(t1,....,tn)(u1,...,um).
This never led to issues until Jore Booy came up with an example
where it did.

Multiple rewriting of this term has been eliminated.

The testcase of Jore has been added.

Some extra asserts have been added to the rewrite_stack.