Resolved bug #1897. Repeated substitution in the compiling rewriter.
If rewriting takes place with a substitution that maps
x to an expression containing itself, the rewriter can sometimes
end in an infinite loop. This is due to substituting x for itself
repeatedly. E.g. with sigma = [x := f(x)] rewriting can cause
the expression with x to be substituted by f(x), and consequently
by f(f(x)), etc.

The rewriter has NOT been designed to not repeatedly apply such
substitutions, and it may still lead to problems if substitutions
in which the substituted variable occurs in the right hand side
are being used.

In case of substituting functions for variables at the head of a
term, e.g. F(1,2) where F is a variable, and F is substitued by
for instance F[x -> y], then repeated subtitution is now
prevented.

This should also mean that rewriting per se of such terms did
become a little more efficient.