Error reported by Jan Friso Groote corrected
There was an error in the following situation:
AvoidLrg finds a new bottom state (because a state with all inert successors
in AvoidLrg has a transition in the large splitter) but NewBotSt is already
aborted.  Then it would not be strictly necessary to add this state to
NewBotSt.  However, when later AvoidLrg is finished, its potential non-bottom
states (not the actual non-bottom states) are added to NewBotSt.  At that
moment, the new bottom state had a counter value that suggested it were in
AvoidLrg proper instead of NewBotSt.

The correction is to add the new bottom state to NewBotSt even if the latter
is already aborted.