Do not generate vertices for X_true and X_false in second pass of pbessolvesymbolic
It turns out we generate many of such vertices, leading to a large overhead and reduced simplification of right hand sides.

Note that this is not an elegant solution. We should rather investigate whether some of the restrictions on the SRF format (in particular requiring it to be recursive) can be dropped, or whether an alternative format such as the parameterized parity game format (PPG) as described by Gijs Kant is more suitable.