Subsequent failed assertion corrected
My previous bugfix led to calling a macro for aborting the NewBotSt coroutine if necessary,
sometimes even when this coroutine has already been aborted earlier.  This is ok, so I
removed the superfluous assertion.

Also formatted the code a bit nicer: generally, the functional code is in columns 1-79
and assertions and other debugging code is in columns 81-999.