Bug around BLC set splitting corrected
Jan Friso Groote found multiple examples of LTSs that led to wrong results in
the experimental algorithm with lazy BLC sets.  The problem was that the condition
to decide whether an empty super-BLC set can be deleted immediately or needs to
be kept (because its place marks where a new main or co-splitter needs to go)
was wrong, leading to the false assumption that some super-BLC sets had no
corresponding main or co-splitter.

On the other hand, when super-BLC sets needed to be split during stabilize(),
there is no need to keep empty super-BLC sets because there is never a main or
co-splitter around.  In this case it happened that the condition was too strict,
leading to spurious empty super-BLC sets that were erroneously not deleted.

The condition is made more explicit and corrected.