Added a few new properties to a self stabilisation protocol.
This protocol is derived from Prism but looks a little different
in the sense that it does not work with rounds in which all parties
are taking place, but the different parties can communicate
asynchronously. This does not affect the correctness, but it
has an effect on the number of steps to reach stabilisation,
which depending on how the actual communications take place
can be fast, in approximately 6 communications for 5 parties,
or slow, requiring on average >22 communications if the order
in which the communications take place is chosen badly.